extract generic execution state
