+++ protostar/commands/test/environments/factory.py
@@ -6,19 +6,19 @@ from protostar.commands.test.environments.setup_execution_environment import (
 from protostar.commands.test.environments.test_execution_environment import (
     TestExecutionEnvironment,
 )
-from protostar.commands.test.execution_state import ExecutionState
 from protostar.commands.test.starkware.execution_resources_summary import (
     ExecutionResourcesSummary,
 )
+from protostar.commands.test.starkware.test_execution_state import TestExecutionState
 
 
-async def invoke_setup(function_name: str, state: ExecutionState):
+async def invoke_setup(function_name: str, state: TestExecutionState):
     env = SetupExecutionEnvironment(state)
     await env.invoke(function_name)
 
 
 async def invoke_test_case(
-    function_name: str, state: ExecutionState
+    function_name: str, state: TestExecutionState
 ) -> Optional[ExecutionResourcesSummary]:
     env = TestExecutionEnvironment(state)
     return await env.invoke(function_name)
