+++ protostar/commands/test/starkware/test_execution_state.py
@@ -1,24 +1,20 @@
 import dataclasses
 from copy import deepcopy
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import List, Optional
 
 from starkware.starknet.services.api.contract_class import ContractClass
-from starkware.starknet.testing.contract import StarknetContract
+from typing_extensions import Self
 
 from protostar.commands.test.test_context import TestContext
+from protostar.starknet.execution_state import ExecutionState
 from protostar.starknet.forkable_starknet import ForkableStarknet
 from protostar.utils.starknet_compilation import StarknetCompiler
 
 
 @dataclass
-class ExecutionState:
-    starknet: ForkableStarknet
-    contract: StarknetContract
-    starknet_compiler: StarknetCompiler
-    include_paths: List[str]
-    disable_hint_validation_in_external_contracts: bool
-    context: TestContext = field(default_factory=TestContext)
+class TestExecutionState(ExecutionState):
+    context: TestContext
 
     @classmethod
     async def from_test_suite_definition(
@@ -27,7 +23,7 @@ class ExecutionState:
         test_suite_definition: ContractClass,
         disable_hint_validation_in_external_contracts: bool,
         include_paths: Optional[List[str]] = None,
-    ) -> "ExecutionState":
+    ) -> Self:
         starknet = await ForkableStarknet.empty()
         contract = await starknet.deploy(contract_class=test_suite_definition)
 
@@ -37,13 +33,8 @@ class ExecutionState:
             starknet_compiler=starknet_compiler,
             include_paths=include_paths or [],
             disable_hint_validation_in_external_contracts=disable_hint_validation_in_external_contracts,
+            context=TestContext(),
         )
 
-    def fork(self) -> "ExecutionState":
-        starknet_fork = self.starknet.fork()
-        return dataclasses.replace(
-            self,
-            starknet=starknet_fork,
-            contract=starknet_fork.copy_and_adapt_contract(self.contract),
-            context=deepcopy(self.context),
-        )
+    def fork(self) -> Self:
+        return dataclasses.replace(super().fork(), context=deepcopy(self.context))
